pyproject.toml
[tool.poetry]
name = "rag-from-pdf"
version = "0.1.0"
description = ""
authors = ["payonear <arthur.kolishenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfminer-six = "^20240706"
ipykernel = "^6.29.5"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
langchain-community = "^0.2.10"
faiss-cpu = "^1.8.0.post1"
tiktoken = "^0.7.0"
pytest = "^8.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
